Add platform tests for Darwin, and Windows
